feat(ui-ux): display block hash and refactor block details
